+++ src/main/java/com/easypost/exception/API/BadRequestError.java
@@ -1,0 +1,20 @@
+package com.easypost.exception.API;
+
+import java.util.List;
+
+import com.easypost.model.Error;
+import com.easypost.exception.APIException;
+
+public class BadRequestError extends APIException {
+    /**
+     * BadRequestError constructor.
+     *
+     * @param message    the exception message
+     * @param code       the exception code
+     * @param statusCode the exception status code
+     * @param errors     the errors array
+     */
+    public BadRequestError(final String message, final String code, final int statusCode, List<Error> errors) {
+        super(message, code, statusCode, errors);
+    }
+}