+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next release
 
 - Adds new `RequestHook` and `ResponseHook` classes. (un)subscribe to them with the new `subscribeToRequestHook`, `subscribeToResponseHook`, `unsubscribeFromRequestHook`, or `unsubscribeFromResponseHook` methods of an `EasyPostClient`
+- Maps 400 status codes to the new `BadRequestError` class
 
 ## v6.7.0 (2023-06-06)